refactor: move graphs elements' views to view.graph
