+++ app/src/main/kotlin/view/graph/GraphView.kt
@@ -1,4 +1,4 @@
-package view
+package view.graph
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxSize
